Move proof primitive tests out of lib
